perf: add suspense to TabName component
