+++ apps/codeimage/tsconfig.json
@@ -22,6 +22,9 @@
       ],
       "@core/*": [
         "./src/core/*"
+      ],
+      "@ui/*": [
+        "./src/ui/*"
       ]
     }
   },
